services/order-service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
kafka-python==2.0.2
pydantic==2.10.4
